# stats.sh: process files in memory instead of on disk