@StevenBlack 2025-08-03
Now can you do [sourcestats.sh](https://github.com/StevenBlack/hosts/blob/master/sourcestats.sh)? 🙂

Merging.