@miallo 2025-08-03
> Now can you do [sourcestats.sh](https://github.com/StevenBlack/hosts/blob/master/sourcestats.sh)? 🙂

Sure thing! (Though today is quite full, so might only be able to do this tomorrow)